01-generation/pyproject.toml
[project]
name = "01-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-aiplatform>=1.140.0",
    "google-cloud-alloydb-connector[pg8000]>=1.12.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-run>=0.15.0",
    "google-genai>=1.65.0",
    "pgvector>=0.4.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "sqlalchemy>=2.0.48",
]
